backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yt-dlp==2023.11.16
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1

